reorganize genome files into one directory
