+++ CMakeLists.txt
@@ -15,12 +15,12 @@ add_executable(
 	src/util.h
 	src/util.cpp
 	src/test.cpp
-	src/phenome/phenome.h
-	src/phenome/phenome.cpp
+        src/genome/phenome.h
+        src/genome/phenome.cpp
 	src/genome/gene/reactionGene.h
 	src/genome/gene/reactionGene.cpp
-	src/phenome/body.cpp
-	src/phenome/body.h
+        src/genome/body.cpp
+        src/genome/body.h
 	src/genome/direction.h
 	src/genome/direction.cpp
 	src/food.h
